C + C4: anyone can tag anyone, and disagree
Revises decision 1, at Benjamin's request — recorded in DECISIONS.md
rather than left as a silent drift from the doc.

The gate is a SIGNER, not a session. isOwner counts a session token as
proof you can sign, which is fine for the prefs flows it gates, but a
token is backend auth and can't sign an event — reusing it here would
render an "Add a tag" button that throws "No signer available" on click.
canTag checks for a local key or window.nostr. Owner vs visitor now
differs only in wording.

Disagreeing is built alongside, because it's the only answer the
protocol has to a bad tag: there is no delete. Re-publishing the same
deterministic d-tag with polarity -1 supersedes the apply — verified on
the relay that the assertion is REPLACED, one event, not stacked. So the
copy says "Disagree", never "Remove"; the old assertion is still on
relays, it just stops counting.

All stance lives in the one popover instead of on the chips. Chips are
links to their tag pages, and hanging a second control off each would
put two tap targets inside a pill on mobile.

Optimistic counts now move our single vote between the applications and
disputes buckets rather than adding a fresh one, and drop a tag that
falls to zero applications — mirroring what the read itself does, so
disagreeing with a tag only you vouched for makes the chip go away and
stay away instead of lingering until the refetch.

Worth stating plainly: the trust filter is currently permissive enough
that any pubkey's tag will show. Disputes contain that, the POV doesn't
yet. Launch gate, not a build gate — same root cause as the own-POV
switch.

Verified with two throwaway keys, tagging each other rather than any
real person: the button appears on someone else's profile, the picker
says "What are they known for?", the tag applies, the picker then offers
Disagree, and disagreeing clears the chip and the relay agrees.

tsc + vite build clean.
